+++ src/core/column-list/column.list.sort.js
@@ -1,22 +1,19 @@
 import { bend, copy } from '../node/node.service';
 import { preOrderDFS } from '../node/node.service';
 
-export { sortIndexFactory, sort };
+export { sortIndexFactory, merge };
 
 function sortIndexFactory(model) {
 	const templateIndex = model.columnList().columns.map(c => c.key);
 
 	return (columns, scores) => {
 		const { length } = columns;
-		scores = Object.assign(
-			{
-				list: column => (column.class === 'data' ? 0.1 : 0.3),
-				index: () => 0.2,
-				view: column => length + (column.class !== 'data' ? 0.1 : 0.3),
-				template: () => length + 0.4
-			},
-			scores
-		);
+		scores = Object.assign({
+			list: column => (column.class === 'data' || column.class === 'cohort') ? 0.1 : 0.3,
+			index: () => 0.2,
+			view: column => length + ((column.class !== 'data' && column.class !== 'cohort') ? 0.1 : 0.3),
+			template: () => length + 0.4
+		}, scores);
 
 		const viewIndex = columns.map(c => c.key);
 
@@ -93,10 +90,11 @@ function equals(xs, ys) {
 			return false;
 		}
 	}
+
 	return true;
 }
 
-function sort(newTree, oldTree, buildIndex) {
+function merge(newTree, oldTree, buildIndex) {
 	const current = running(newTree, buildIndex);
 	const screen = former(oldTree, current);
 	const insertNear = insertFactory(current, screen);
@@ -129,12 +127,12 @@ function sort(newTree, oldTree, buildIndex) {
 function running(tree, buildIndex) {
 	const result = {
 		line: [],
-		set: new Set()
+		map: new Map()
 	};
 
 	preOrderDFS([tree], node => {
 		result.line.push(node);
-		result.set.add(node.key.model.key);
+		result.map.set(node.key.model.key, node.key);
 
 		// As we use pre order direction we can manipulate with children without affecting on algorithm.
 		// Below we sort columns in appropriate order.
@@ -162,8 +160,11 @@ function former(tree, current) {
 	preOrderDFS([tree], node => {
 		// Filter out nodes if they were deleted from newTree.
 		const { key } = node.key.model;
-		if (current.set.has(key)) {
-			result.line.push(copy(node));
+		const view = current.map.get(key);
+		if (view) {
+			const newNode = copy(node);
+			newNode.key = view;
+			result.line.push(newNode);
 			result.set.add(key);
 		}
 	});