+++ src/lib/main/column/column.component.ts
@@ -1,8 +1,9 @@
-import { Component, Input, OnInit, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
+import { Component, Input, ChangeDetectionStrategy, OnDestroy, SkipSelf, Optional, OnInit, ElementRef } from '@angular/core';
 import { isUndefined } from 'ng2-qgrid/core/utility/kit';
 import { ColumnModel } from 'ng2-qgrid/core/column-type/column.model';
 import { TemplateHostService } from '../../template/template-host.service';
 import { ColumnListService } from '../../main/column/column-list.service';
+import { guid } from 'ng2-qgrid/core/services/guid';
 import { ColumnService } from './column.service';
 
 @Component({
@@ -12,11 +13,9 @@ import { ColumnService } from './column.service';
 	changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ColumnComponent implements OnInit, OnDestroy {
-	model: ColumnModel;
-
 	@Input() type: string;
 	@Input() key: string;
-	@Input() class: 'data' | 'control' | 'markup' | 'pivot';
+	@Input() class: 'data' | 'control' | 'markup' | 'pivot' | 'cohort';
 	@Input() title: string;
 	@Input() pin: null | 'left' | 'right';
 	@Input() aggregation: string;
@@ -65,58 +64,74 @@ export class ColumnComponent implements OnInit, OnDestroy {
 	constructor(
 		private columnList: ColumnListService,
 		private templateHost: TemplateHostService,
-		private columnService: ColumnService
+		@SkipSelf() @Optional() private parent: ColumnService,
+		private service: ColumnService,
+		private element: ElementRef
 	) { }
 
 	ngOnInit() {
-		const withKey = !isUndefined(this.key);
-		const withType = !isUndefined(this.type);
+		let withKey = !isUndefined(this.key);
+		let withType = !isUndefined(this.type);
+
+		// We want to update model when ngOntInit is triggered and not in afterViewInit
+		// so we apply dirty hack to understand if column is cohort or not.
+		const element = this.element.nativeElement as HTMLElement;
+		if (element.children.length && element.children.item(0).tagName === 'Q-GRID-COLUMN') {
+			this.type = 'cohort';
+			if (!withKey) {
+				this.key = `$cohort-${this.title || guid()}`;
+			}
+
+			withKey = true;
+			withType = true;
+		}
+
 		if (!withKey) {
 			this.key = this.columnList.generateKey(this);
 		}
 
 		const column = this.columnList.extract(this.key, this.type);
+		this.columnList.copy(column, this);
 
 		this.templateHost.key = source => {
 			const parts = [source, 'cell'];
 
 			if (withType) {
-				parts.push(column.type);
+				parts.push(this.type);
 			}
 
 			if (withKey) {
-				parts.push(`the-${column.key}`);
+				parts.push(`the-${this.key}`);
 			}
 
 			return parts.join('-') + '.tpl.html';
 		};
 
-		this.columnList.copy(column, this);
-
-		this.columnService.column = column;
-		const { parent } = this.columnService;
-		if (withKey || parent.parent) {
-			this.columnList.add(column, parent.column);
+		if (withKey) {
+			if (this.parent) {
+				this.parent.column.children.push(column);
+			} else {
+				this.columnList.add(column);
+			}
+			this.service.column = column;
 		} else {
-			const settings = Object.keys(this)
-				.filter(
-					key => !isUndefined(this[key]) && column.hasOwnProperty(key)
-				)
-				.reduce((memo, key) => {
-					memo[key] = column[key];
-					return memo;
-				}, {});
+			const settings =
+				Object
+					.keys(this)
+					.filter(key => !isUndefined(this[key]) && column.hasOwnProperty(key))
+					.reduce((memo, key) => {
+						memo[key] = column[key];
+						return memo;
+					}, {});
 
 			this.columnList.register(settings);
 		}
-
-		this.model = column;
 	}
 
 	ngOnDestroy() {
-		const { model } = this;
-		if (model && model.source === 'template') {
-			this.columnList.delete(model.key);
+		const { column } = this.service;
+		if (column && column.source === 'template') {
+			this.columnList.delete(column.key);
 		}
 	}
 }