+++ src/lib/plugins/validation/validation.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input } from '@angular/core';
+import { PluginService } from '../plugin.service';
+
+@Component({
+	selector: 'q-grid-validation',
+	template: '',
+	providers: [PluginService]
+})
+export class ValidationComponent {
+	@Input() type: string;
+
+	constructor(private plugin: PluginService) {
+	}
+}