+++ src/lib/template/template-host.service.ts
@@ -3,5 +3,5 @@ import { identity } from 'ng2-qgrid/core/utility/kit';
 
 @Injectable()
 export class TemplateHostService {
-	public key: (x: string) => string = identity;
+	key: (x: string) => string = identity;
 }